fix archived's import path
